request trace: keep runtime entry points in module root
